refactor(go-template): extract value-lowering into value/value-lowering.ts
Move the signal/const initial-value baking cluster out of the adapter
class into pure free functions over GoEmitContext (Roadmap B). The
adapter object keeps only orchestration; value lowering becomes a
standalone, stub-testable module.

- value/value-lowering.ts: convertInitialValue, jsLiteralToGo,
  objectLiteralToGoMap, tsLiteralToGo, getSignalInitialValueAsGo.
- emit-context.ts: GoEmitContext gains typeInfoToGo and
  extractPropNameFromInitialValue (the entry points the moved
  functions call back into); typeInfoToGo / parseLiteralExpression
  stay on the adapter as widely-shared members.

Output byte-identical, verified by the adapter unit (552) and
adapter-tests conformance (786) suites; tsgo + biome clean.
